+++ config/level/nikic/php-parser40.yml
@@ -1,10 +1,10 @@
 services:
-    Rector\Rector\Contrib\PhpParser\IdentifierRector: ~
-    Rector\Rector\Contrib\PhpParser\ParamAndStaticVarNameRector: ~
-    Rector\Rector\Contrib\PhpParser\CatchAndClosureUseNameRector: ~
-    Rector\Rector\Contrib\PhpParser\SetLineRector: ~
-    Rector\Rector\Contrib\PhpParser\RemoveNodeRector: ~
-    Rector\Rector\Contrib\PhpParser\UseWithAliasRector: ~
+    Rector\PhpParser\Rector\IdentifierRector: ~
+    Rector\PhpParser\Rector\ParamAndStaticVarNameRector: ~
+    Rector\PhpParser\Rector\CatchAndClosureUseNameRector: ~
+    Rector\PhpParser\Rector\SetLineRector: ~
+    Rector\PhpParser\Rector\RemoveNodeRector: ~
+    Rector\PhpParser\Rector\UseWithAliasRector: ~
 
     Rector\Rector\Dynamic\PropertyNameReplacerRector:
         $perClassOldToNewProperties:
