+++ packages/PhpParser/src/Rector/ParamAndStaticVarNameRector.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Rector\Contrib\PhpParser;
+namespace Rector\PhpParser\Rector;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\PropertyFetch;
