+++ packages/PhpParser/src/config/services.yml
@@ -1,0 +1,7 @@
+services:
+    _defaults:
+        autowire: true
+
+    Rector\PhpParser\:
+        resource: '../'
+        exclude: '../{Rector/**/*Rector.php}'
