+++ packages/PhpParser/tests/Rector/CatchAndClosureUseNameRector/Correct/correct.php.inc
