+++ packages/PhpParser/tests/Rector/CatchAndClosureUseNameRector/Wrong/wrong.php.inc
