+++ packages/PhpParser/tests/Rector/CatchAndClosureUseNameRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\CatchAndClosureUseNameRector: ~
