+++ packages/PhpParser/tests/Rector/IdentifierRector/Correct/correct.php.inc
