+++ packages/PhpParser/tests/Rector/IdentifierRector/Correct/correct2.php.inc
