+++ packages/PhpParser/tests/Rector/IdentifierRector/Correct/correct3.php.inc
