+++ packages/PhpParser/tests/Rector/IdentifierRector/Correct/correct4.php.inc
