+++ packages/PhpParser/tests/Rector/IdentifierRector/Wrong/wrong.php.inc
