+++ packages/PhpParser/tests/Rector/IdentifierRector/Wrong/wrong2.php.inc
