+++ packages/PhpParser/tests/Rector/IdentifierRector/Wrong/wrong3.php.inc
