+++ packages/PhpParser/tests/Rector/IdentifierRector/Wrong/wrong4.php.inc
