+++ packages/PhpParser/tests/Rector/IdentifierRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\IdentifierRector: ~
