+++ packages/PhpParser/tests/Rector/ParamAndStaticVarNameRector/Correct/correct.php.inc
