+++ packages/PhpParser/tests/Rector/ParamAndStaticVarNameRector/ParamAndStaticVarNameRectorTest.php
@@ -1,14 +1,14 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Contrib\PhpParser\RemoveNodeRector;
+namespace Rector\PhpParser\Tests\Rector\ParamAndStaticVarNameRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
 /**
- * @covers \Rector\Rector\Contrib\PhpParser\RemoveNodeRector
+ * @covers \Rector\PhpParser\Rector\ParamAndStaticVarNameRector
  */
-final class RemoveNodeRectorTest extends AbstractRectorTestCase
+final class ParamAndStaticVarNameRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -21,7 +21,6 @@ final class RemoveNodeRectorTest extends AbstractRectorTestCase
     public function provideWrongToFixedFiles(): Iterator
     {
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
     }
 
     protected function provideConfig(): string
