+++ packages/PhpParser/tests/Rector/ParamAndStaticVarNameRector/Wrong/wrong.php.inc
