+++ packages/PhpParser/tests/Rector/ParamAndStaticVarNameRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\ParamAndStaticVarNameRector: ~
