+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Correct/correct.php.inc
