+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Correct/correct2.php.inc
