+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Correct/correct3.php.inc
