+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Wrong/wrong.php.inc
