+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Wrong/wrong2.php.inc
