+++ packages/PhpParser/tests/Rector/RemoveNodeRector/Wrong/wrong3.php.inc
