+++ packages/PhpParser/tests/Rector/RemoveNodeRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\RemoveNodeRector: ~
