+++ packages/PhpParser/tests/Rector/SetLineRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\SetLineRector: ~
