+++ packages/PhpParser/tests/Rector/UseWithAliasRector/Correct/correct.php.inc
