+++ packages/PhpParser/tests/Rector/UseWithAliasRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\PhpParser\Rector\UseWithAliasRector: ~
