--- tests/Rector/Contrib/PhpParser/CatchAndClosureUseNameRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\CatchAndClosureUseNameRector: ~
