--- tests/Rector/Contrib/PhpParser/IdentifierRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\IdentifierRector: ~
