--- tests/Rector/Contrib/PhpParser/ParamAndStaticVarNameRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\ParamAndStaticVarNameRector: ~
