--- tests/Rector/Contrib/PhpParser/RemoveNodeRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\RemoveNodeRector: ~
