--- tests/Rector/Contrib/PhpParser/SetLineRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\SetLineRector: ~
