--- tests/Rector/Contrib/PhpParser/UseWithAliasRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Rector\Contrib\PhpParser\UseWithAliasRector: ~
